+++ src/main/java/com/webapps2022/entity/SystemUser.java
@@ -5,6 +5,7 @@
 package com.webapps2022.entity;
 
 import java.io.Serializable;
+import java.math.BigDecimal;
 import java.util.List;
 import java.util.Objects;
 import javax.persistence.Entity;
@@ -39,7 +40,7 @@ public class SystemUser implements Serializable {
     String surname;
 
     @NotNull
-    Float currencyCount;
+    BigDecimal currencyCount;
 
     @NotNull
     CurrencyEnum currencyType;
@@ -57,7 +58,7 @@ public class SystemUser implements Serializable {
     public SystemUser() {
     }
 
-    public SystemUser(String username, String password, String name, String surname, Float currencyCount,
+    public SystemUser(String username, String password, String name, String surname, BigDecimal currencyCount,
             CurrencyEnum currencyType, SystemUserGroup group) {
         this.username = username;
         this.password = password;
@@ -100,11 +101,11 @@ public class SystemUser implements Serializable {
         this.surname = surname;
     }
 
-    public Float getCurrencyCount() {
+    public BigDecimal getCurrencyCount() {
         return currencyCount;
     }
 
-    public void setCurrencyCount(Float currencyCount) {
+    public void setCurrencyCount(BigDecimal currencyCount) {
         this.currencyCount = currencyCount;
     }
